+++ src/bitap-search/constants.js
@@ -1,0 +1,2 @@
+// Machine word size
+module.exports.MAX_BITS = 32