+++ src/extended-search/exact-match.js
@@ -8,10 +8,12 @@ const sanitize = pattern => pattern.substr(1)
 
 const match = (pattern, text) => {
   const sanitizedPattern = sanitize(pattern)
-  const isMatch = text.indexOf(sanitizedPattern) > -1
+  const index = text.indexOf(sanitizedPattern)
+  const isMatch = index > -1
+
   return {
     isMatch,
-    score: 0
+    score: 0,
   }
 }