+++ src/extended-search/prefix-exact-match.js
@@ -9,6 +9,7 @@ const sanitize = pattern => pattern.substr(1)
 const match = (pattern, text) => {
   const sanitizedPattern = sanitize(pattern)
   const isMatch = text.startsWith(sanitizedPattern)
+
   return {
     isMatch,
     score: 0