+++ src/ngram-search/array-utils/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  union: require('./union'),
+  intersection: require('./intersection')
+}