+++ src/ngram-search/distance/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  jaccardDistance: require('./jaccard-distance'),
+  tverskyIndex: require('./tversky-index')
+}